+++ packages/core-js-pure/override/internals/set-helpers.js
@@ -9,7 +9,5 @@ module.exports = {
   add: caller('add', 1),
   has: caller('has', 1),
   remove: caller('delete', 1),
-  proto: SetPrototype,
-  $has: SetPrototype.has,
-  $keys: SetPrototype.keys
+  proto: SetPrototype
 };