+++ tests/compat/tests.js
@@ -1714,7 +1714,7 @@ GLOBAL.tests = {
     return Set.from;
   },
   'esnext.set.intersection.v2': function () {
-    return Set.prototype.intersection;
+    return Array.from(new Set([1, 2, 3]).intersection(new Set([3, 2]))) == '3,2';
   },
   'esnext.set.is-disjoint-from.v2': function () {
     return Set.prototype.isDisjointFrom;